feat: Add print functionality and improve UI
Integrates the `react-to-print` library to allow users to print defect reports directly from the application. This enhances the workflow for documenting and sharing product quality issues.

Additionally, this commit refactors the DraggableFAB component, moves it to a dedicated file, and introduces a new Header component to improve overall application structure and user interface organization. Minor UI adjustments are also made to the DashboardReport for better layout and visual appeal.